Revert "@Erisa requested changes."
This reverts commit 4a454e81758c211464cd940923081277de931886.
