+++ Program.cs
@@ -215,6 +215,7 @@ namespace Cliptok
                     badMsgLog = await discord.GetChannelAsync(cfgjson.InvestigationsChannelId);
                     Mutes.CheckMutesAsync();
                     ModCmds.CheckBansAsync();
+                    MemberCmds.CheckRemindersAsync();
 
                     string commitHash = "";
                     string commitMessage = "";
@@ -582,6 +583,7 @@ namespace Cliptok
             commands.RegisterCommands<ModCmds>();
             commands.RegisterCommands<Lockdown>();
             commands.RegisterCommands<Bans>();
+            commands.RegisterCommands<MemberCmds>();
             commands.CommandErrored += CommandsNextService_CommandErrored;
 
             await discord.ConnectAsync();
@@ -591,9 +593,11 @@ namespace Cliptok
                 await Task.Delay(10000);
                 Mutes.CheckMutesAsync();
                 ModCmds.CheckBansAsync();
-                ModCmds.CheckRemindersAsync();
+                MemberCmds.CheckRemindersAsync();
 #pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
+
             }
+
         }
     }
 }
